\name{Nest}
\Rdversion{2.1}
\alias{Nest}
\docType{data}
\title{
Nest orientations and creek directions
}
\description{Orientation of nests of 50 noisy scrub birds (theta) along the bank of a creek bed, together
with the corresponding directions (phi) of creek flow at the nearest point to the nest.
}
\usage{data(Nest)}
\format{Orientation of 50 nests (vectors)}
\source{
Data supplied by Dr. Graham Smith}
\references{
N.I. Fisher (1993) Statistical analysis of circular data. Cambridge University Press.}

\keyword{datasets}
