\name{WindDirectionsTrivariate}
\Rdversion{2.1}
\alias{WindDirectionsTrivariate}
\docType{data}
\title{Wind directions}
\description{Wind directions registered at the monitoring stations of San Agustin located in the north,
Pedregal in the southwest, and Hangares in the southeast of the Mexico Central Valley's at 14:00 on days
between January 1, 1993 and February 29, 2000. There are a total of 1,682 observations}
\usage{data(WindDirectionsTrivariate)}
\format{Three columns of angles in radians}
\source{Mexico Central Valleys pollution monitoring network.
RAMA SIMAT (Red Automatica de Monitoreo Ambiental) }
\keyword{datasets}
